Nearly finish SVG files and fix Google font plug
This commit will fix the SVGs used in the about section. I've also recreated the rest of the SVGs in public/img/svg/ - they're ready for title tags and acessibility info.

I'm mainly pushing this commit to test my added media queries, because "responsive" Firefox inspector doesn't respond to them... Or maybe it does and I'm dumb - hence the test!